requirements-dev.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=/home/jameslp/autograder/contrib/dev_scripts/../requirements-dev.txt /home/jameslp/autograder/contrib/dev_scripts/../requirements-dev.in
#
aiohttp==3.9.1
    # via black
aiosignal==1.3.1
    # via aiohttp
annotated-types==0.6.0
    # via
    #   -c /home/jameslp/autograder/contrib/dev_scripts/../requirements.txt
    #   pydantic
argcomplete==3.2.1
    # via datamodel-code-generator
async-timeout==4.0.3
    # via aiohttp
attrs==23.2.0
    # via aiohttp
black==24.1a1
    # via
    #   -r /home/jameslp/autograder/contrib/dev_scripts/../requirements-dev.in
    #   datamodel-code-generator
click==8.1.7
    # via black
datamodel-code-generator==0.25.2
    # via -r /home/jameslp/autograder/contrib/dev_scripts/../requirements-dev.in
dnspython==2.4.2
    # via
    #   -c /home/jameslp/autograder/contrib/dev_scripts/../requirements.txt
    #   email-validator
email-validator==2.1.0.post1
    # via
    #   -c /home/jameslp/autograder/contrib/dev_scripts/../requirements.txt
    #   pydantic
exceptiongroup==1.2.2
    # via pytest
execnet==2.1.1
    # via pytest-xdist
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
genson==1.2.2
    # via datamodel-code-generator
idna==3.6
    # via
    #   -c /home/jameslp/autograder/contrib/dev_scripts/../requirements.txt
    #   email-validator
    #   yarl
inflect==5.6.2
    # via datamodel-code-generator
iniconfig==2.0.0
    # via pytest
isort==5.13.2
    # via
    #   -r /home/jameslp/autograder/contrib/dev_scripts/../requirements-dev.in
    #   datamodel-code-generator
jinja2==3.1.3
    # via datamodel-code-generator
markupsafe==2.1.4
    # via jinja2
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.0.0
    # via black
nodeenv==1.8.0
    # via pyright
packaging==23.2
    # via
    #   black
    #   datamodel-code-generator
    #   pytest
pathspec==0.12.1
    # via black
platformdirs==4.1.0
    # via black
pluggy==1.5.0
    # via pytest
psutil==6.1.0
    # via pytest-xdist
pycodestyle==2.11.1
    # via -r /home/jameslp/autograder/contrib/dev_scripts/../requirements-dev.in
pydantic[email]==2.5.3
    # via
    #   -c /home/jameslp/autograder/contrib/dev_scripts/../requirements.txt
    #   datamodel-code-generator
pydantic-core==2.14.6
    # via
    #   -c /home/jameslp/autograder/contrib/dev_scripts/../requirements.txt
    #   pydantic
pydocstyle==6.3.0
    # via -r /home/jameslp/autograder/contrib/dev_scripts/../requirements-dev.in
pyright==1.1.347
    # via -r /home/jameslp/autograder/contrib/dev_scripts/../requirements-dev.in
pytest==8.3.3
    # via
    #   -r /home/jameslp/autograder/contrib/dev_scripts/../requirements-dev.in
    #   pytest-xdist
pytest-xdist[psutil]==3.6.1
    # via -r /home/jameslp/autograder/contrib/dev_scripts/../requirements-dev.in
pyyaml==6.0.1
    # via
    #   -c /home/jameslp/autograder/contrib/dev_scripts/../requirements.txt
    #   datamodel-code-generator
snowballstemmer==2.2.0
    # via pydocstyle
toml==0.10.2
    # via datamodel-code-generator
tomli==2.0.1
    # via
    #   black
    #   pytest
types-pyyaml==6.0.12.12
    # via -r /home/jameslp/autograder/contrib/dev_scripts/../requirements-dev.in
types-requests==2.31.0.20240106
    # via -r /home/jameslp/autograder/contrib/dev_scripts/../requirements-dev.in
typing-extensions==4.9.0
    # via
    #   -c /home/jameslp/autograder/contrib/dev_scripts/../requirements.txt
    #   black
    #   pydantic
    #   pydantic-core
urllib3==2.1.0
    # via
    #   -c /home/jameslp/autograder/contrib/dev_scripts/../requirements.txt
    #   types-requests
yarl==1.9.4
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# setuptools
